--- backend/db/example_config.json
@@ -1,5 +1,0 @@
-{
-  "token": "",
-  "owner": "google",
-  "repo": "go-github"
-}
