+++ backend/db_pkg/.gitignore
@@ -1,0 +1,1 @@
+*.json
