+++ backend/web_pkg/util.go
@@ -1,15 +1,6 @@
-package main
+package web_pkg
 
-import (
-	"context"
-	"encoding/json"
-	"log"
-	"net/http"
-	"os"
-
-	"github.com/google/go-github/v68/github"
-	"golang.org/x/oauth2"
-)
+import "net/http"
 
 func setHeaders(w *http.ResponseWriter, content_type string) {
 
@@ -23,7 +14,7 @@ func setHeaders(w *http.ResponseWriter, content_type string) {
 	(*w).Header().Set("Access-Control-Allow-Headers", "Content-Type")
 }
 
-func enable_cors(handler http.Handler) http.Handler {
+func EnableCors(handler http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 
 		w.Header().Set("Access-Control-Allow-Origin", "*")
@@ -39,37 +30,4 @@ func enable_cors(handler http.Handler) http.Handler {
 		// Serve the actual request
 		handler.ServeHTTP(w, r)
 	})
-}
-
-func load_config() *Config {
-	content, err := os.ReadFile("./db/config.json")
-	if err != nil {
-		log.Fatal("Error when opening config: ", err)
-	}
-
-	var payload Config
-	err = json.Unmarshal(content, &payload)
-	if err != nil {
-		log.Fatal("Error during Unmarshal of config: ", err)
-	}
-
-	return &payload
-}
-
-func init_github_connection(ctx context.Context) (*Config, *github.Client) {
-	config := load_config()
-
-	authToken := config.Token
-
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: authToken},
-	)
-	tc := oauth2.NewClient(ctx, ts)
-
-	client := github.NewClient(tc)
-
-	config.Token = ""
-
-	return config, client
-
 }
