+++ package.json
@@ -1,15 +1,15 @@
 {
-  "name": "gh-pull-request-dashboard",
-  "version": "2024.0",
-  "scripts": {
-    "start": "concurrently \"npm run dev --prefix frontend\" \"cd backend && go run .\"",
-    "lin": "concurrently \"npm run preview --prefix frontend\" \"cd backend && .\\gh-pull-request-dashboard\"",
-    "win": "concurrently \"npm run preview --prefix frontend\" \"cd backend && .\\gh-pull-request-dashboard.exe\"",
-    "build": "npm run build --prefix frontend",
-    "dev": "npm run dev --prefix frontend",
-    "backend": "cd backend && go run ."
-  },
-  "devDependencies": {
-    "concurrently": "^9.0.1"
-  }
+	"name": "gh-pull-request-dashboard",
+	"version": "2024.0",
+	"scripts": {
+		"start": "concurrently \"npm run dev --prefix frontend\" \"cd backend && go run .\"",
+		"lin": "concurrently \"npm run preview --prefix frontend\" \"cd backend && .\\github-pull-request-dashboard\"",
+		"win": "concurrently \"npm run preview --prefix frontend\" \"cd backend && .\\github-pull-request-dashboard.exe\"",
+		"build": "npm run build --prefix frontend",
+		"dev": "npm run dev --prefix frontend",
+		"backend": "cd backend && go run ."
+	},
+	"devDependencies": {
+		"concurrently": "^9.0.1"
+	}
 }
